requirements.txt
# Core Backend Dependencies
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-JWT-Extended==4.7.1
Flask-Caching==2.3.1
Flask-Limiter==3.5.0
Flask-CORS==5.0.0
python-dotenv==1.2.1

# Database
PyMySQL==1.1.2
cryptography==44.0.1

# Serialization & Validation
marshmallow==3.20.1
marshmallow-sqlalchemy==1.4.2

# Utilities
requests==2.32.5
pathlib==1.0.1

